makeSpecializationSet: use templated isl types
Use explicit API calls instead of operators because the use of aff_set
is confusing (at least to the author of this commit).
The conversion to an isl::PwAff is needed because, unlike the Python
bindings, the C++ bindings do not have a class hierarchy.
The conversion could also be avoided by introducing
an isl_aff_zero_set in isl.