prefixScheduleMupa: use templated isl types
An explicit cast is introduced in prefixScheduleMupa,
which will be removed when infixScheduleMupa is converted
to templated isl types.
Automatic type deduction is temporarily turned off in
MappedScop::detectReductions to avoid having to convert
the entire MappedScop::detectReductions together with prefixScheduleMupa.
It will be reintroduced which will be removed when memberRange is converted
to templated isl types.
Automatic type deduction is also temporarily turned off in
partialScheduleMupa.  It will be reintroduced
after the schedule tree has been converted to templated isl types.